Move tunnel provider to the root
